(PC-34274)[API] refactor: centralize user eligibility functions
