+++ api/src/pcapi/core/users/models.py
@@ -389,9 +389,9 @@ class User(PcObject, Base, Model, DeactivableMixin):
 
     @property
     def eligibility(self) -> EligibilityType | None:
-        from pcapi.core.fraud import api as fraud_api
+        from pcapi.core.users import eligibility_api
 
-        return fraud_api.decide_eligibility(self, self.birth_date, datetime.utcnow())
+        return eligibility_api.decide_eligibility(self, self.birth_date, datetime.utcnow())
 
     @hybrid_property
     def full_name(self) -> str:
